separate non-rDSN-dependent utilities from dev.cpp into dev.utility to avoid circular dependency (not compile on linux)
