+++ bin/dsn.cmake
@@ -379,7 +379,7 @@ function(dsn_add_project)
             if(DSN_BUILD_RUNTIME AND(DEFINED DSN_IN_CORE) AND DSN_IN_CORE)
                 set(TEMP_LIBS "")
             else()
-                set(TEMP_LIBS dsn.dev.cpp dsn.core)
+                set(TEMP_LIBS dsn.dev.cpp dsn.core dsn.dev.utility)
             endif()
             set(MY_PROJ_LIBS ${MY_PROJ_LIBS} ${MY_BOOST_LIBS} ${TEMP_LIBS})
             if(${CMAKE_SYSTEM_NAME} STREQUAL "Linux")
