+++ include/dsn/c/api_layer1.h
@@ -942,6 +942,15 @@ typedef struct
     int size;
 } dsn_file_buffer_t;
 
+/*! the following ctrl code are used by \ref dsn_file_ctrl. */
+typedef enum dsn_ctrl_code_t
+{
+    CTL_BATCH_INVALID = 0,
+    CTL_BATCH_WRITE = 1,            ///< (batch) set write batch size
+    CTL_MAX_CON_READ_OP_COUNT = 2,  ///< (throttling) maximum concurrent read ops
+    CTL_MAX_CON_WRITE_OP_COUNT = 3, ///< (throttling) maximum concurrent write ops
+} dsn_ctrl_code_t;
+
 /*!
  open file
