+++ include/dsn/cpp/json_helper.h
@@ -45,7 +45,7 @@
 #include <cctype>
 #include <dsn/utility/autoref_ptr.h>
 #include <dsn/cpp/auto_codes.h>
-#include <dsn/cpp/utils.h>
+#include <dsn/utility/utils.h>
 #include <dsn/dist/replication/replication_types.h>
 #include <dsn/dist/replication/replication_other_types.h>
