+++ include/dsn/cpp/perf_test_helper.h
@@ -36,7 +36,7 @@
 # pragma once
 
 # include <dsn/service_api_cpp.h>
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include <sstream>
 # include <atomic>
 # include <vector>
