+++ include/dsn/cpp/serialization.h
@@ -35,7 +35,7 @@
 
 # pragma once
 
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include <dsn/cpp/address.h>
 # include <dsn/cpp/rpc_stream.h>
