+++ include/dsn/cpp/serialization_helper/protobuf_helper.h
@@ -36,7 +36,7 @@
 # pragma once
 
 # include <dsn/utility/ports.h>
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 
 # include <google/protobuf/stubs/common.h>
 # include <google/protobuf/io/zero_copy_stream.h>
