+++ include/dsn/cpp/task_helper.h
@@ -37,7 +37,7 @@
 
 # include <dsn/service_api_c.h>
 # include <dsn/cpp/auto_codes.h>
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include <dsn/cpp/rpc_stream.h>
 # include <dsn/cpp/serialization.h>
 # include <dsn/cpp/zlocks.h>
