+++ include/dsn/tool-api/nfs.h
@@ -37,7 +37,7 @@
 
 # include <dsn/service_api_c.h>
 # include <string>
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include <dsn/tool-api/task.h>
 
 namespace dsn {
