+++ src/core/core/disk_engine.cpp
@@ -36,7 +36,7 @@
 # include "disk_engine.h"
 # include <dsn/tool-api/perf_counters.h>
 # include <dsn/tool-api/aio_provider.h>
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include "transient_memory.h"
 
 # ifdef __TITLE__
