+++ src/core/core/env_provider.cpp
@@ -34,7 +34,7 @@
  */
 
 # include <dsn/tool-api/env_provider.h>
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include <chrono>
 
 namespace dsn {
