+++ src/core/core/library_utils.cpp
@@ -33,7 +33,7 @@
  *     xxxx-xx-xx, author, fix bug about xxx
  */
 
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include <dsn/utility/singleton.h>
 # include <sys/types.h>
 # include <sys/stat.h>
