+++ src/core/tests/file_utils.cpp
@@ -36,7 +36,7 @@
 
 # include <gtest/gtest.h>
 # include <dsn/service_api_cpp.h>
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include <fstream>
 
 static void file_utils_test_setup()
