+++ src/core/tests/service_api_c.cpp
@@ -37,7 +37,7 @@
 # include <dsn/tool_api.h>
 # include <dsn/tool-api/task.h>
 # include <dsn/cpp/auto_codes.h>
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include <gtest/gtest.h>
 # include <thread>
 # include "../core/service_engine.h"
