+++ src/core/tests/sim_lock.cpp
@@ -37,7 +37,7 @@
 # include <dsn/tool_api.h>
 # include <dsn/tool-api/task.h>
 # include <dsn/cpp/auto_codes.h>
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include <dsn/utility/synchronize.h>
 # include <gtest/gtest.h>
 # include <thread>
