+++ src/core/tests/utils.cpp
@@ -34,7 +34,7 @@
  */
 
 
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include <dsn/utility/link.h>
 # include <dsn/utility/autoref_ptr.h>
 # include <gtest/gtest.h>
