+++ src/core/tools/hpc/hpc_tail_logger.cpp
@@ -36,7 +36,7 @@
 
 # include "hpc_tail_logger.h"
 # include <dsn/utility/singleton_store.h>
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include <dsn/tool-api/command.h>
 # include <cstdlib>
 # include <sstream>
