+++ src/dev/CMakeLists.txt
@@ -1,1 +1,2 @@
 add_subdirectory(cpp)
+add_subdirectory(utility)
