+++ src/dev/cpp/file_utils.cpp
@@ -34,7 +34,7 @@
  */
 
 
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include <sys/stat.h>
 # include <errno.h>
