+++ src/tools/repli/repli.main.cpp
@@ -34,7 +34,7 @@
  */
 
 # include "repli.app.h"
-# include <dsn/cpp/utils.h>
+# include <dsn/utility/utils.h>
 # include <iostream>
 # include <thread>
 # if !defined (_WIN32)
